feat(pox-4-tests): add check function delegate to PoxCommand-derived types
- Separate success paths from failure paths to keep pox_Commands.ts focused
  on success cases only. This prevents the file from growing with
  out-of-scope cases.

Note: This is a work in progress.
