+++ contrib/boot-contracts-stateful-prop-tests/tests/pox-4/pox_Commands.ts
@@ -1,11 +1,9 @@
 import fc from "fast-check";
-import { Real, Stacker, Stub, StxAddress, Wallet } from "./pox_CommandModel";
+import { PoxCommand, Stacker, StxAddress, Wallet } from "./pox_CommandModel";
 import { GetStackingMinimumCommand } from "./pox_GetStackingMinimumCommand";
 import { GetStxAccountCommand } from "./pox_GetStxAccountCommand";
 import { StackStxSigCommand } from "./pox_StackStxSigCommand";
-import { StackStxSigCommand_Err } from "./pox_StackStxSigCommand_Err";
 import { StackStxAuthCommand } from "./pox_StackStxAuthCommand";
-import { StackStxAuthCommand_Err } from "./pox_StackStxAuthCommand_Err";
 import { DelegateStxCommand } from "./pox_DelegateStxCommand";
 import { DelegateStackStxCommand } from "./pox_DelegateStackStxCommand";
 import { Simnet } from "@hirosystems/clarinet-sdk";
@@ -29,7 +27,7 @@ export function PoxCommands(
   wallets: Map<StxAddress, Wallet>,
   stackers: Map<StxAddress, Stacker>,
   network: Simnet,
-): fc.Arbitrary<Iterable<fc.Command<Stub, Real>>> {
+): fc.Arbitrary<PoxCommand>[] {
   const cmds = [
     // GetStackingMinimumCommand
     fc.record({
@@ -85,36 +83,6 @@ export function PoxCommands(
         r.margin,
       )
     ),
-    // StackStxAuthCommand_Err
-    fc.record({
-      wallet: fc.constantFrom(...wallets.values()),
-      authId: fc.nat(),
-      period: fc.integer({ min: 1, max: 12 }),
-      margin: fc.integer({ min: 1, max: 9 }),
-    }).map((
-      r: {
-        wallet: Wallet;
-        authId: number;
-        period: number;
-        margin: number;
-      },
-    ) =>
-      new StackStxAuthCommand_Err(
-        r.wallet,
-        r.authId,
-        r.period,
-        r.margin,
-        function (this: StackStxAuthCommand_Err, model: Readonly<Stub>): boolean {
-          const stacker = model.stackers.get(this.wallet.stxAddress)!;
-          console.log("I in StackStxAuthCommand_Err stacker", stacker);
-          return (
-            model.stackingMinimum > 0 && !stacker.isStacking &&
-            !stacker.hasDelegated
-          );
-        },
-        123,
-      )
-    ),
     // StackExtendAuthCommand
     fc
       .record({
@@ -137,36 +105,6 @@ export function PoxCommands(
             r.currentCycle,
           ),
       ),
-    // StackStxSigCommand_Err
-    fc.record({
-      wallet: fc.constantFrom(...wallets.values()),
-      authId: fc.nat(),
-      period: fc.integer({ min: 1, max: 12 }),
-      margin: fc.integer({ min: 1, max: 9 }),
-    }).map((
-      r: {
-        wallet: Wallet;
-        authId: number;
-        period: number;
-        margin: number;
-      },
-    ) =>
-      new StackStxSigCommand_Err(
-        r.wallet,
-        r.authId,
-        r.period,
-        r.margin,
-        function (this: StackStxSigCommand_Err, model: Readonly<Stub>): boolean {
-          const stacker = model.stackers.get(this.wallet.stxAddress)!;
-          console.log("I in StackStxSigCommand_Err stacker", stacker);
-          return (
-            model.stackingMinimum > 0 && !stacker.isStacking &&
-            !stacker.hasDelegated
-          );
-        },
-        123,
-      )
-    ),
     // StackExtendSigCommand
     fc
       .record({
@@ -514,9 +452,7 @@ export function PoxCommands(
     ),
   ];
 
-  // More on size: https://github.com/dubzzz/fast-check/discussions/2978
-  // More on cmds: https://github.com/dubzzz/fast-check/discussions/3026
-  return fc.commands(cmds, { size: "xsmall" });
+  return cmds;
 }
 
 export const REWARD_CYCLE_LENGTH = 1050;
